Cadastro de partido corrigido com o banco de dados.

+++ src/br/com/etec/view/jinternalframe/TelaCadastroPartido.java
@@ -7,6 +7,7 @@ package br.com.etec.view.jinternalframe;
 
 import br.com.etec.dao.PartidoDao;
 import br.com.etec.model.Partido;
+import br.com.etec.utils.DbUtils;
 import br.com.etec.utils.ManipularImagem;
 import java.awt.Color;
 import java.awt.Container;
@@ -16,6 +17,8 @@ import java.awt.event.ActionListener;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.text.ParseException;
 import java.util.logging.Level;
@@ -39,8 +42,6 @@ public class TelaCadastroPartido extends JInternalFrame {
     private JTextField txtPartido;
     private JLabel lblSlogan;
     private JTextField txtSlogan;
-    private JLabel lblDtCriacao;
-    private JFormattedTextField txtDtCriacao;
 
     public TelaCadastroPartido() throws ParseException {
         iniciandoCompomentes();
@@ -58,73 +59,56 @@ public class TelaCadastroPartido extends JInternalFrame {
         JLabel lblCamposObri = new JLabel("(*)Campos Obrigatórios");
         lblCamposObri.setBounds(600, 20, 200, 20);
         lblCamposObri.setForeground(Color.red);
-        //Id
-        lblId = new JLabel("*ID");
-        lblId.setBounds(200, 120, 80, 25);
-        lblId.setForeground(Color.black);
-
-        txtId = new JTextField();
-        txtId.setBounds(255, 120, 100, 25);
-        txtId.setEditable(false);
 
         // Nome do Candidato
         lblNome = new JLabel("*Nome");
-        lblNome.setBounds(200, 150, 200, 25);
+        lblNome.setBounds(150, 150, 200, 25);
         lblNome.setForeground(Color.black);
 
         txtNome = new JTextField(10);
-        txtNome.setBounds(255, 150, 200, 25);
+        txtNome.setBounds(205, 150, 200, 25);
 
         // Partido
         lblPartido = new JLabel("*Sigla");
-        lblPartido.setBounds(460, 150, 80, 25);
+        lblPartido.setBounds(415, 120, 80, 25);
         lblPartido.setForeground(Color.black);
 
         txtPartido = new JTextField();
-        txtPartido.setBounds(505, 150, 60, 25);
+        txtPartido.setBounds(410, 150, 60, 25);
 
         // Número
         lblNumero = new JLabel("*Número");
-        lblNumero.setBounds(200, 180, 80, 25);
+        lblNumero.setBounds(150, 180, 80, 25);
         lblNumero.setForeground(Color.black);
 
         txtNumero = new JFormattedTextField(new MaskFormatter("##"));
-        txtNumero.setBounds(255, 180, 30, 25);
+        txtNumero.setBounds(205, 180, 30, 25);
 
-        //Cargo
-        lblSlogan = new JLabel(" Slogan");
-        lblSlogan.setBounds(200, 210, 80, 25);
+        //Sloga
+        lblSlogan = new JLabel("Slogan");
+        lblSlogan.setBounds(150, 210, 80, 25);
         lblSlogan.setForeground(Color.black);
 
         txtSlogan = new JTextField();
-        txtSlogan.setBounds(255, 210, 200, 25);
+        txtSlogan.setBounds(205, 210, 200, 25);
 
         //Imagem
-        lblFoto = new JLabel(" Logo");
-        lblFoto.setBounds(200, 240, 80, 25);
+        lblFoto = new JLabel("Logo");
+        lblFoto.setBounds(150, 240, 80, 25);
         lblFoto.setForeground(Color.black);
 
         txtFoto = new JTextField(10);
-        txtFoto.setBounds(255, 240, 200, 25);
+        txtFoto.setBounds(205, 240, 200, 25);
         txtFoto.setEnabled(false);
 
-        lblImagem = new JLabel();
-        lblImagem.setBounds(550, 240, 200, 200);
+        ImageIcon imgLogo = new ImageIcon(getClass().getResource("/br/com/etec/imgs/logoUserBD.png"));
+        lblImagem = new JLabel(imgLogo);
+        lblImagem.setToolTipText("Logo");
+        lblImagem.setBounds(520, 100, 200, 200);
 
         btnEnviar = new JButton("Enviar");
-        btnEnviar.setBounds(460, 240, 80, 25);
+        btnEnviar.setBounds(410, 240, 80, 25);
 
-        //
-        lblDtCriacao = new JLabel(" Data de Criação");
-        lblDtCriacao.setBounds(160, 270, 120, 25);
-        lblDtCriacao.setForeground(Color.black);
-
-        MaskFormatter mask = new MaskFormatter();
-        mask.setMask("##/##/####");
-        mask.setPlaceholderCharacter('_');
-
-        txtDtCriacao = new JFormattedTextField(mask);
-        txtDtCriacao.setBounds(255, 270, 200, 25);
         // Buttons
         ImageIcon imgAdicionar = new ImageIcon(getClass().getResource("/br/com/etec/imgs/add.png"));
         btnAdicionar = new JButton(imgAdicionar);
@@ -134,18 +118,17 @@ public class TelaCadastroPartido extends JInternalFrame {
             @Override
             public void actionPerformed(ActionEvent e) {
                 Partido addPartido = new Partido();
-                if ((txtNome.getText().isEmpty() || txtNumero.getText().isEmpty() || txtDtCriacao.getText().isEmpty() || txtPartido.getText().isEmpty())) {
+                if ((txtNome.getText().isEmpty() || txtNumero.getText().isEmpty() || txtPartido.getText().isEmpty())) {
                     JOptionPane.showMessageDialog(null, "Todos os campos (*) obrigatórios");
-                } else if ((txtNumero.getText().length() < 2) || (txtDtCriacao.getText().length() < 10)) {
+                } else if ((txtNumero.getText().length() < 2)) {
                     JOptionPane.showMessageDialog(null, "Preencha todos os campos corretamente");
                 } else {
                     try {
-                        addPartido.setId(1);
+                        addPartido.setIdNumero(Integer.parseInt(txtNumero.getText()));
+                        System.err.println("Erro: " + txtNumero.getText());
                         addPartido.setNome(txtNome.getText());
-                        addPartido.setNumero(Integer.parseInt(txtNumero.getText()));
                         addPartido.setSigla(txtPartido.getText());
                         addPartido.setSlogan(txtSlogan.getText());
-                        addPartido.setData(txtDtCriacao.getText());
                         if (imagem != null) {
                             addPartido.setLogo(ManipularImagem.getImgBytes(imagem));
                         }
@@ -159,6 +142,9 @@ public class TelaCadastroPartido extends JInternalFrame {
                         JOptionPane.showMessageDialog(null, "Erro ao criar partido" + ex.getMessage());
                     } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | SQLException ex) {
                         Logger.getLogger(TelaCadastroCandidato.class.getName()).log(Level.SEVERE, null, ex);
+                        if (ex instanceof com.mysql.jdbc.exceptions.jdbc4.MySQLIntegrityConstraintViolationException) {
+                            JOptionPane.showMessageDialog(null, "Partido já cadastrado");
+                        }
                     }
                 }
             }
@@ -180,18 +166,16 @@ public class TelaCadastroPartido extends JInternalFrame {
             public void actionPerformed(ActionEvent e
             ) {
                 Partido addPartido = new Partido();
-                if ((txtNome.getText().isEmpty() || txtNumero.getText().isEmpty() || txtDtCriacao.getText().isEmpty() || txtPartido.getText().isEmpty())) {
+                if ((txtNome.getText().isEmpty() || txtNumero.getText().isEmpty() || txtPartido.getText().isEmpty())) {
                     JOptionPane.showMessageDialog(null, "Todos os campos (*) obrigatórios");
-                } else if ((txtNumero.getText().length() < 2) || (txtDtCriacao.getText().length() < 10)) {
+                } else if ((txtNumero.getText().length() < 2)) {
                     JOptionPane.showMessageDialog(null, "Preencha todos os campos corretamente");
                 } else {
                     try {
-                        addPartido.setId(1);
+                        addPartido.setIdNumero(Integer.parseInt(txtNumero.getText()));
                         addPartido.setNome(txtNome.getText());
-                        addPartido.setNumero(Integer.parseInt(txtNumero.getText()));
                         addPartido.setSigla(txtPartido.getText());
                         addPartido.setSlogan(txtSlogan.getText());
-                        addPartido.setData(txtDtCriacao.getText());
                         if (imagem != null) {
                             addPartido.setLogo(ManipularImagem.getImgBytes(imagem));
                         }
@@ -225,24 +209,18 @@ public class TelaCadastroPartido extends JInternalFrame {
             @Override
             public void actionPerformed(ActionEvent e
             ) {
-
-                int numero;
                 Partido excPart = new Partido();
-
+                excPart.setIdNumero(Integer.parseInt(txtNumero.getText()));
                 try {
-                    numero = Integer.parseInt(JOptionPane.showInputDialog("Número do Partido"));
-                    excPart.setNumero(numero);
-                } catch (NumberFormatException nfe) {
-                    JOptionPane.showMessageDialog(null, "Apenas números");
-                }
-
-                try {
-                    new PartidoDao().delete(excPart);
-                    desabilita();
-                    clearCampos();
-                    JOptionPane.showMessageDialog(null, "Partido excluído com sucesso!");
-                } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | SQLException ex) {
-                    Logger.getLogger(TelaCadastroCandidato.class.getName()).log(Level.SEVERE, null, ex);
+                    int confir = JOptionPane.showConfirmDialog(null, "Deseja excluir", "Atenção", JOptionPane.YES_NO_CANCEL_OPTION);
+                    if (confir == JOptionPane.YES_OPTION) {
+                        new PartidoDao().delete(excPart);
+                        JOptionPane.showMessageDialog(null, "Usuário deletado com sucesso!");
+                        clearCampos();
+                        desabilita();
+                    }
+                } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | SQLException | HeadlessException ex) {
+                    JOptionPane.showMessageDialog(null, "Erro ao cadastrar usuário");
                 }
             }
         }
@@ -266,22 +244,23 @@ public class TelaCadastroPartido extends JInternalFrame {
                     int numero = 0;
                     try {
                         numero = Integer.parseInt(JOptionPane.showInputDialog("Número do Partido"));
+                        partido = new PartidoDao().findById(numero);
+
+                            System.err.println(partido);
+                            txtNome.setText(partido.getNome());
+                            txtNumero.setText("" + partido.getIdNumero());
+                            txtPartido.setText(partido.getSigla());
+                            txtSlogan.setText(partido.getSlogan());
+
+                            ManipularImagem.exibirImagemLabel(partido.getLogo(), lblImagem);
+
+                            habilita();
+                        
                     } catch (NumberFormatException nfe) {
                         JOptionPane.showMessageDialog(null, "Apenas números");
                     }
-                    partido = new PartidoDao().findById(numero);
 
-                    txtId.setText(String.valueOf(partido.getId()));
-                    txtNome.setText(partido.getNome());
-                    txtNumero.setText("" + partido.getNumero());
-                    txtPartido.setText(partido.getSigla());
-                    txtSlogan.setText(partido.getSlogan());
-                    txtDtCriacao.setText(partido.getData());
-                    ManipularImagem.exibirImagemLabel(partido.getLogo(), lblImagem);
-
-                    habilita();
-
-                } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | SQLException | HeadlessException ex) {
+                } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | SQLException | HeadlessException | NullPointerException ex) {
                     JOptionPane.showMessageDialog(null, "Erro ao pesquisar" + ex.getMessage());
                 }
             }
@@ -322,47 +301,30 @@ public class TelaCadastroPartido extends JInternalFrame {
         //Add
         container.add(lblCamposObri);
 
-        container.add(lblDtCriacao);
-
-        container.add(txtDtCriacao);
-
+        //container.add(lblDtCriacao);
+        //container.add(txtDtCriacao);
         container.add(lblNome);
-
         container.add(txtNome);
 
         container.add(lblNumero);
-
         container.add(txtNumero);
 
         container.add(lblPartido);
+        container.add(txtPartido);
 
         container.add(lblFoto);
-
         container.add(txtFoto);
 
         container.add(btnEnviar);
-
-        container.add(lblId);
-
-        container.add(txtId);
-
         container.add(btnAdicionar);
-
         container.add(btnAtualizar);
-
         container.add(btnExcluir);
-
         container.add(btnPesquisar);
 
         container.add(lblImagem);
 
-        container.add(txtPartido);
-
         container.add(lblSlogan);
-
         container.add(txtSlogan);
-
-        container.add(txtNumero);
 
         setClosable(
                 true);
@@ -373,26 +335,26 @@ public class TelaCadastroPartido extends JInternalFrame {
     }
 
     public void clearCampos() {
-        txtId.setText(null);
         txtNome.setText(null);
         txtFoto.setText(null);
         txtNumero.setValue(null);
-        lblImagem.setIcon(null);
+        lblImagem.setIcon(new ImageIcon(getClass().getResource("/br/com/etec/imgs/logoUserBD.png")));
         txtPartido.setText(null);
         txtSlogan.setText(null);
-        txtDtCriacao.setText(null);
     }
 
     // Habilita os botões de Excluir e atualizar
     public void habilita() {
         btnAtualizar.setEnabled(true);
         btnExcluir.setEnabled(true);
+        txtNumero.setEditable(false);
     }
 
     // Desabilita os botões de Excluir e atualizar
     public void desabilita() {
         btnAtualizar.setEnabled(false);
         btnExcluir.setEnabled(false);
+        txtNumero.setEditable(true);
     }
 
     private JButton btnAdicionar;
@@ -405,13 +367,11 @@ public class TelaCadastroPartido extends JInternalFrame {
     private JFormattedTextField txtNumero;
     private JLabel lblPartido;
     private JLabel lblFoto;
-    BufferedImage imagem;
+    private BufferedImage imagem;
     private JButton btnEnviar;
-    private JLabel lblId;
-    private JTextField txtId;
     private JLabel lblImagem;
     private Partido partido;
     private JTextField txtFoto;
-    Connection connection;
+    private Connection connection;
 
 }
